# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: YuTong Platform API
  description: "YuTong é\x9B¨æ¡\x90 - å\x95\x86ä¸\x9Açº§å\N¨æ \x88æ\x8A\x80æ\x9C¯åº\x95\
    åº§ API å¥\x91çº¦"
  contact:
    name: YuTong Team
  version: 0.2.0
servers:
- url: http://localhost:8082
  description: Generated server url
security:
- bearerAuth: []
paths:
...
  /api/v1/ai/prompt-templates/{id}:
    get:
      tags:
      - "AI-Promptæ¨¡æ\x9D¿ç®¡ç\x90\x86"
      summary: "æ\x9F¥è¯¢æ¨¡æ\x9D¿è¯¦æ\x83\N"
      operationId: getAiPromptTemplate
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultAiPromptTemplate'
    put:
      tags:
      - "AI-Promptæ¨¡æ\x9D¿ç®¡ç\x90\x86"
      summary: "æ\x9B´æ\x96°æ¨¡æ\x9D¿"
      operationId: updateAiPromptTemplate
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AiPromptTemplate'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultAiPromptTemplate'
  /api/v1/ai/knowledge-bases/{id}:
    get:
      tags:
      - "AI-ç\x9F¥è¯\x86åº\x93ç®¡ç\x90\x86"
      summary: "æ\x9F¥è¯¢ç\x9F¥è¯\x86åº\x93è¯¦æ\x83\N"
      operationId: getAiKnowledgeBase
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultAiKnowledgeBase'
    put:
      tags:
      - "AI-ç\x9F¥è¯\x86åº\x93ç®¡ç\x90\x86"
      summary: "æ\x9B´æ\x96°ç\x9F¥è¯\x86åº\x93è\x8D\x89ç¨¿"
      operationId: updateAiKnowledgeBase
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AiKnowledgeBase'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultAiKnowledgeBase'
...
  /api/v1/ai/prompt-templates:
    get:
      tags:
      - "AI-Promptæ¨¡æ\x9D¿ç®¡ç\x90\x86"
      summary: "å\x88\x86é¡µæ\x9F¥è¯¢ Prompt æ¨¡æ\x9D¿"
      operationId: listAiPromptTemplates
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 1
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 20
      - name: templateCode
        in: query
        required: false
        schema:
          type: string
      - name: scenario
        in: query
        required: false
        schema:
          type: string
      - name: status
        in: query
        required: false
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultPageResultAiPromptTemplate'
    post:
      tags:
      - "AI-Promptæ¨¡æ\x9D¿ç®¡ç\x90\x86"
      summary: "å\x88\x9Bå»ºæ¨¡æ\x9D¿è\x8D\x89ç¨¿"
      operationId: createAiPromptTemplate
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AiPromptTemplate'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultAiPromptTemplate'
  /api/v1/ai/prompt-templates/{id}/publish:
    post:
      tags:
      - "AI-Promptæ¨¡æ\x9D¿ç®¡ç\x90\x86"
      summary: "å\x8F\x91å¸\x83æ¨¡æ\x9D¿"
      operationId: publishAiPromptTemplate
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: version
        in: query
        required: true
        schema:
          type: integer
          format: int32
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultAiPromptTemplate'
  /api/v1/ai/knowledge-bases:
    get:
      tags:
      - "AI-ç\x9F¥è¯\x86åº\x93ç®¡ç\x90\x86"
      summary: "å\x88\x86é¡µæ\x9F¥è¯¢ç\x9F¥è¯\x86åº\x93"
      operationId: listAiKnowledgeBases
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 1
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 20
      - name: kbCode
        in: query
        required: false
        schema:
          type: string
      - name: status
        in: query
        required: false
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultPageResultAiKnowledgeBase'
    post:
      tags:
      - "AI-ç\x9F¥è¯\x86åº\x93ç®¡ç\x90\x86"
      summary: "å\x88\x9Bå»ºç\x9F¥è¯\x86åº\x93è\x8D\x89ç¨¿"
      operationId: createAiKnowledgeBase
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AiKnowledgeBase'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultAiKnowledgeBase'
  /api/v1/ai/knowledge-bases/{id}/publish:
    post:
      tags:
      - "AI-ç\x9F¥è¯\x86åº\x93ç®¡ç\x90\x86"
      summary: "å\x8F\x91å¸\x83ç\x9F¥è¯\x86åº\x93"
      operationId: publishAiKnowledgeBase
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: version
        in: query
        required: true
        schema:
          type: integer
          format: int32
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultAiKnowledgeBase'
  /api/v1/ai/knowledge-bases/{id}/disable:
    post:
      tags:
      - "AI-ç\x9F¥è¯\x86åº\x93ç®¡ç\x90\x86"
      summary: "ç¦\x81ç\x94¨ç\x9F¥è¯\x86åº\x93"
      operationId: disableAiKnowledgeBase
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: version
        in: query
        required: true
        schema:
          type: integer
          format: int32
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultAiKnowledgeBase'
  /api/v1/ai/documents/{id}/mark-stale:
    post:
      tags:
      - "AI-æ\x96\x87æ¡£å\N¥åº\x93ç®¡ç\x90\x86"
      summary: "æ \x87è®°æ\x96\x87æ¡£å¤±æ\x95\x88"
      operationId: markAiDocumentStale
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: version
        in: query
        required: true
        schema:
          type: integer
          format: int32
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultAiDocument'
  /api/v1/ai/documents/ingest:
    post:
      tags:
      - "AI-æ\x96\x87æ¡£å\N¥åº\x93ç®¡ç\x90\x86"
      summary: "å\x8F\x97æ\x8E§å\N¥åº\x93æ\x96\x87æ¡£"
      operationId: ingestAiDocument
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IngestDocumentRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultAiDocument'
...
  /api/v1/ai/documents:
    get:
      tags:
      - "AI-æ\x96\x87æ¡£å\N¥åº\x93ç®¡ç\x90\x86"
      summary: "å\x88\x86é¡µæ\x9F¥è¯¢æ\x96\x87æ¡£"
      operationId: listAiDocuments
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 1
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 20
      - name: kbId
        in: query
        required: false
        schema:
          type: string
      - name: docTitle
        in: query
        required: false
        schema:
          type: string
      - name: documentStatus
        in: query
        required: false
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultPageResultAiDocument'
  /api/v1/ai/documents/{id}:
    get:
      tags:
      - "AI-æ\x96\x87æ¡£å\N¥åº\x93ç®¡ç\x90\x86"
      summary: "æ\x9F¥è¯¢æ\x96\x87æ¡£è¯¦æ\x83\N"
      operationId: getAiDocument
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ResultAiDocument'
...
components:
  schemas:
...
    AiPromptTemplate:
      type: object
      properties:
        id:
          type: string
        tenantId:
          type: string
        createdBy:
          type: string
        createdTime:
          type: string
          format: date-time
        updatedBy:
          type: string
        updatedTime:
          type: string
          format: date-time
        deleted:
          type: boolean
        version:
          type: integer
          format: int32
        remark:
          type: string
        templateCode:
          type: string
        versionNo:
          type: integer
          format: int32
        scenario:
          type: string
        inputSchema:
          type: string
        outputSchema:
          type: string
        promptText:
          type: string
        safetyRules:
          type: string
        evaluationSetCode:
          type: string
        status:
          type: string
    ResultAiPromptTemplate:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        messageKey:
          type: string
        messageArgs:
          type: object
          additionalProperties: {}
        data:
          $ref: '#/components/schemas/AiPromptTemplate'
        traceId:
          type: string
        timestamp:
          type: string
          format: date-time
    AiKnowledgeBase:
      type: object
      properties:
        id:
          type: string
        tenantId:
          type: string
        createdBy:
          type: string
        createdTime:
          type: string
          format: date-time
        updatedBy:
          type: string
        updatedTime:
          type: string
          format: date-time
        deleted:
          type: boolean
        version:
          type: integer
          format: int32
        remark:
          type: string
        kbCode:
          type: string
        kbName:
          type: string
        description:
          type: string
        aclPolicyJson:
          type: string
        embeddingModel:
          type: string
        visibility:
          type: string
        permissionCode:
          type: string
        sensitivityLevel:
          type: string
        status:
          type: string
        ownerUserId:
          type: string
    ResultAiKnowledgeBase:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        messageKey:
          type: string
        messageArgs:
          type: object
          additionalProperties: {}
        data:
          $ref: '#/components/schemas/AiKnowledgeBase'
        traceId:
          type: string
        timestamp:
          type: string
          format: date-time
...
    AiDocument:
      type: object
      properties:
        id:
          type: string
        tenantId:
          type: string
        createdBy:
          type: string
        createdTime:
          type: string
          format: date-time
        updatedBy:
          type: string
        updatedTime:
          type: string
          format: date-time
        deleted:
          type: boolean
        version:
          type: integer
          format: int32
        remark:
          type: string
        kbId:
          type: string
        fileId:
          type: string
        docTitle:
          type: string
        sourceType:
          type: string
        sourceUri:
          type: string
        visibility:
          type: string
        permissionCode:
          type: string
        sensitivityLevel:
          type: string
        documentStatus:
          type: string
        chunkCount:
          type: integer
          format: int32
        errorMessage:
          type: string
        indexedTime:
          type: string
          format: date-time
    ResultAiDocument:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        messageKey:
          type: string
        messageArgs:
          type: object
          additionalProperties: {}
        data:
          $ref: '#/components/schemas/AiDocument'
        traceId:
          type: string
        timestamp:
          type: string
          format: date-time
    IngestDocumentRequest:
      type: object
      properties:
        kbId:
          type: string
        docTitle:
          type: string
        sourceType:
          type: string
        sourceUri:
          type: string
        content:
          type: string
        visibility:
          type: string
        sensitivityLevel:
          type: string
        permissionCode:
          type: string
...
    PageResultAiPromptTemplate:
      type: object
      properties:
        records:
          type: array
          items:
            $ref: '#/components/schemas/AiPromptTemplate'
        total:
          type: integer
          format: int64
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
    ResultPageResultAiPromptTemplate:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        messageKey:
          type: string
        messageArgs:
          type: object
          additionalProperties: {}
        data:
          $ref: '#/components/schemas/PageResultAiPromptTemplate'
        traceId:
          type: string
        timestamp:
          type: string
          format: date-time
    PageResultAiKnowledgeBase:
      type: object
      properties:
        records:
          type: array
          items:
            $ref: '#/components/schemas/AiKnowledgeBase'
        total:
          type: integer
          format: int64
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
    ResultPageResultAiKnowledgeBase:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        messageKey:
          type: string
        messageArgs:
          type: object
          additionalProperties: {}
        data:
          $ref: '#/components/schemas/PageResultAiKnowledgeBase'
        traceId:
          type: string
        timestamp:
          type: string
          format: date-time
    PageResultAiDocument:
      type: object
      properties:
        records:
          type: array
          items:
            $ref: '#/components/schemas/AiDocument'
        total:
          type: integer
          format: int64
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
    ResultPageResultAiDocument:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        messageKey:
          type: string
        messageArgs:
          type: object
          additionalProperties: {}
        data:
          $ref: '#/components/schemas/PageResultAiDocument'
        traceId:
          type: string
        timestamp:
          type: string
          format: date-time
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: MOCK